Move event handling for alerts to a separate class
