+++ version.properties
@@ -1,5 +1,5 @@
-#Sat Dec 03 04:06:41 EST 2022
-VERSION_BUILD=1441
+#Sat Dec 03 10:10:43 EST 2022
+VERSION_BUILD=1442
 VERSION_PHASE=release
 VERSION_MAJOR=1
 VERSION_MINOR=3
